+++ CHANGELOG.md
@@ -1,9 +1,33 @@
-## [2.193.1](https://github.com/bettyblocks/material-ui-component-set/compare/v2.193.0...v2.193.1) (2024-10-07)
+## [2.193.2](https://github.com/bettyblocks/material-ui-component-set/compare/v2.193.1...v2.193.2) (2024-10-11)
 
 
 ### Bug Fixes
 
+* add default value to current filter and show remaining records ([2a51f91](https://github.com/bettyblocks/material-ui-component-set/commit/2a51f919d97cc3c189b7259d3a6d9deda4b833bc))
+* datepicker can interpret default meta-data data ([c56ecb3](https://github.com/bettyblocks/material-ui-component-set/commit/c56ecb39c78bf70050d2647cd3c6c0f71f8f0f1e))
+* deepcopy optionFilter in autocomplete ([b777209](https://github.com/bettyblocks/material-ui-component-set/commit/b777209a4b011df97391741d2476ee82f7b99323))
+* fixed filter in the autocomplete ([a6c71d3](https://github.com/bettyblocks/material-ui-component-set/commit/a6c71d313b0516e08a701c61920e6f359f15a160))
+* force floatingLabel when a value is present in dataPicker ([d2997db](https://github.com/bettyblocks/material-ui-component-set/commit/d2997db89d5fe9a794186bfdc7ff205ce4873638))
+* give datePickerField a default value of null ([028183e](https://github.com/bettyblocks/material-ui-component-set/commit/028183e2456b05dbebe9ca908c29f54988a8df06))
+* parse min- and maxDate correctly ([4baa79e](https://github.com/bettyblocks/material-ui-component-set/commit/4baa79e6a2a64de49253f08d4193674c6025cd58))
+* pipelines ([d72e9ec](https://github.com/bettyblocks/material-ui-component-set/commit/d72e9ec380e289999528d109b857505a841c9336))
+* use empty string as fallback ([0a2d6a9](https://github.com/bettyblocks/material-ui-component-set/commit/0a2d6a9c22d855ee65c3f91913c9a78ac099ba33))
+
+## [2.193.1](https://github.com/bettyblocks/material-ui-component-set/compare/v2.193.0...v2.193.1) (2024-10-11)
+
+
+### Bug Fixes
+
+* add default value to current filter and show remaining records ([2a51f91](https://github.com/bettyblocks/material-ui-component-set/commit/2a51f919d97cc3c189b7259d3a6d9deda4b833bc))
+* datepicker can interpret default meta-data data ([c56ecb3](https://github.com/bettyblocks/material-ui-component-set/commit/c56ecb39c78bf70050d2647cd3c6c0f71f8f0f1e))
+* deepcopy optionFilter in autocomplete ([b777209](https://github.com/bettyblocks/material-ui-component-set/commit/b777209a4b011df97391741d2476ee82f7b99323))
+* fixed filter in the autocomplete ([a6c71d3](https://github.com/bettyblocks/material-ui-component-set/commit/a6c71d313b0516e08a701c61920e6f359f15a160))
+* force floatingLabel when a value is present in dataPicker ([d2997db](https://github.com/bettyblocks/material-ui-component-set/commit/d2997db89d5fe9a794186bfdc7ff205ce4873638))
+* give datePickerField a default value of null ([028183e](https://github.com/bettyblocks/material-ui-component-set/commit/028183e2456b05dbebe9ca908c29f54988a8df06))
 * **PAGE-4710:** Remove unused zone from publish prod files. ([05e1bbf](https://github.com/bettyblocks/material-ui-component-set/commit/05e1bbfe41028b96f0960f51e61c5f8c860bc23e))
+* parse min- and maxDate correctly ([4baa79e](https://github.com/bettyblocks/material-ui-component-set/commit/4baa79e6a2a64de49253f08d4193674c6025cd58))
+* pipelines ([d72e9ec](https://github.com/bettyblocks/material-ui-component-set/commit/d72e9ec380e289999528d109b857505a841c9336))
+* use empty string as fallback ([0a2d6a9](https://github.com/bettyblocks/material-ui-component-set/commit/0a2d6a9c22d855ee65c3f91913c9a78ac099ba33))
 
 # [2.193.0](https://github.com/bettyblocks/material-ui-component-set/compare/v2.192.0...v2.193.0) (2024-10-04)